+++ CHANGES_NEXT_RELEASE
@@ -6,16 +6,16 @@
   * #1573:  pagination not working when entities have the exact same creation date (sort also by entity id)
   * #1583:  If a 'q' contains an ampersand, that ampersand needs to be forwarded as a semicolon
   * #1593:  Modify the type of an entity in a Replace operation (only working in legacy mode)
+  * Fixed a possible crash for TRoE and attributes of type Vocab/Json/Language
+  * Forbidden to DELETE the Core Context !!!  (it can be reloaded)
+  * Bug fix: JSON NULL literal is no longer forwarded!
 
 ## New Features:
   * Support for attributes of type VocabularyProperty
   * Support for attributes of type JsonProperty
   * Support for the new URL parameter "format" for output formats (normalized, concise, simplified)
-  * Support for JsonProperty attribute type
-  * Fixed a possible crash for TRoE and attributes of type Vocab/Json/Language
-  * Forbidden to DELETE the Core Context !!!  (it can be reloaded)
-  * Bug fix: JSON NULL literal is no longer forwarded!
-
+  * New service: DELETE /ngsi-ld/v1/entities  (URL param 'type' only - the rest are missing still)
+  
 ## Notes
   * TRoE is still not prepared for attributes of type Vocab/Json/Language, so, attributes of those types are not stored in the historical database
   * Modified the @context hosting feature to be according to API spec
