+++ test/functionalTest/cases/0000_ngsild/ngsild_delete_entities.test
@@ -25,9 +25,10 @@ Entity Batch Delete Test
 
 --SHELL-INIT--
 dbInit CB
-orionldStart CB
+dbInit CP1
+orionldStart CB -experimental -forwarding
+orionldStart CP1 -experimental
 accumulatorStart --pretty-print 127.0.0.1 ${LISTENER_PORT}
-
 
 --SHELL--
 
@@ -36,8 +37,12 @@ accumulatorStart --pretty-print 127.0.0.1 ${LISTENER_PORT}
 # 02. Create a subscription on entity type T1
 # 03. Remove entities E1 and E3 using DELETE /ngsi-ld/v1/entities?type=T1
 # 04. Get all entities - see E2 and E4
-# 05. Dump accumulator, see ONE notification with TWO deleted entities
+# 05. Dump/Reset accumulator, see ONE notification with TWO deleted entities
 # 06. Attempt to remove entities with type T1 again - see 204
+# 07. Create entity urn:E5, of type T2, in CP1
+# 08. Create a registration for entity type T2, pointing to CP1
+# 09. Remove entities E2 and E4 using DELETE /ngsi-ld/v1/entities?type=T2 - provoke forwarded request to CP1
+# 
 #
 
 echo "01. Create 4 entities: E1, E2, E3, and E4, odd ids with type T1, even with type T2"
@@ -120,8 +125,8 @@ echo
 echo
 
 
-echo "05. Dump accumulator, see ONE notification with TWO deleted entities"
-echo "===================================================================="
+echo "05. Dump/Reset accumulator, see ONE notification with TWO deleted entities"
+echo "=========================================================================="
 accumulatorDump
 echo
 echo
@@ -133,6 +138,63 @@ orionCurl --url /ngsi-ld/v1/entities?type=T1 -X DELETE
 echo
 echo
 
+
+echo "07. Create entity urn:E5, of type T2, in CP1"
+echo "============================================"
+payload='{
+  "id": "urn:E5",
+  "type": "T2",
+  "A": {
+    "type": "Property",
+    "value": 5
+  }
+}'
+orionCurl --url /ngsi-ld/v1/entities --payload "$payload" --port $CP1_PORT
+echo
+echo
+
+
+echo "08. Create a registration for entity type T2, pointing to CP1"
+echo "============================================================="
+payload='{
+  "id": "urn:R1",
+  "type": "ContextSourceRegistration",
+  "information": [
+    {
+      "entities": [
+        {
+          "type": "T2"
+        }
+      ]
+    }
+  ],
+  "endpoint": "http://'$HOSTNAME':'$CP1_PORT'",
+  "operations": [ "purgeEntity" ]
+}'
+orionCurl --url /ngsi-ld/v1/csourceRegistrations --payload "$payload"
+echo
+echo
+
+
+echo "09. Remove entities E2 and E4 using DELETE /ngsi-ld/v1/entities?type=T2 - provoke forwarded request to CP1"
+echo "=========================================================================================================="
+orionCurl --url /ngsi-ld/v1/entities?type=T2 -X DELETE
+echo
+echo
+
+
+echo "10. GET all entities locally in CB - see empty array"
+echo "===================================================="
+orionCurl --url "/ngsi-ld/v1/entities?options=keyValues&local=true"
+echo
+echo
+
+
+echo "11. GET all entities locally in CP1 - see empty array"
+echo "====================================================="
+orionCurl --url "/ngsi-ld/v1/entities?options=keyValues&local=true" --port $CP1_PORT
+echo
+echo
 
 
 --REGEXPECT--
@@ -198,8 +260,8 @@ Link: <https://uri.etsi.org/ngsi-ld/v1/ngsi-ld-core-contextREGEX(.*)
 ]
 
 
-05. Dump accumulator, see ONE notification with TWO deleted entities
-====================================================================
+05. Dump/Reset accumulator, see ONE notification with TWO deleted entities
+==========================================================================
 POST http://REGEX(.*)/notify?subscriptionId=urn:S1
 Content-Length: 331
 User-Agent: orionld/REGEX(.*)
@@ -235,6 +297,51 @@ Ngsild-Attribute-Format: Normalized
 HTTP/1.1 204 No Content
 Date: REGEX(.*)
 
+
+
+07. Create entity urn:E5, of type T2, in CP1
+============================================
+HTTP/1.1 201 Created
+Content-Length: 0
+Date: REGEX(.*)
+Location: /ngsi-ld/v1/entities/urn:E5
+
+
+
+08. Create a registration for entity type T2, pointing to CP1
+=============================================================
+HTTP/1.1 201 Created
+Content-Length: 0
+Date: REGEX(.*)
+Location: /ngsi-ld/v1/csourceRegistrations/urn:R1
+
+
+
+09. Remove entities E2 and E4 using DELETE /ngsi-ld/v1/entities?type=T2 - provoke forwarded request to CP1
+==========================================================================================================
+HTTP/1.1 204 No Content
+Date: REGEX(.*)
+
+
+
+10. GET all entities locally in CB - see empty array
+====================================================
+HTTP/1.1 200 OK
+Content-Length: 2
+Content-Type: application/json
+Date: REGEX(.*)
+
+[]
+
+
+11. GET all entities locally in CP1 - see empty array
+=====================================================
+HTTP/1.1 200 OK
+Content-Length: 2
+Content-Type: application/json
+Date: REGEX(.*)
+
+[]
 
 
 --TEARDOWN--
