+++ test/functionalTest/cases/0000_ngsild/ngsild_options-method.test
@@ -228,7 +228,7 @@ echo
 01. OPTIONS request for /entities
 =================================
 HTTP/1.1 200 OK
-Allow: GET,POST,OPTIONS
+Allow: GET,POST,DELETE,OPTIONS
 Content-Length: 0
 Date: REGEX(.*)
